Twitch/YouTube live stream work
Got the initial code down for the bot to start looking at YouTube and Twitch streaming on the servers the bot is included on. Don't know if it works now I don't have a fake streaming account to use to test the bot.

Made all strings standardized to '' instead of using both '' and ""